requirements.txt
asgiref==3.8.1
Django==5.1.4
paho-mqtt==2.1.0
sqlparse==0.5.2
